Update index.html with new images and adjust image heights
This commit updates the index.html file to include new images (cablecar.jpg, city.jpg, and plane.jpg) and sets a consistent height of 150 pixels for all images. The "Setting image width and height" paragraph was also removed, and an exclamation mark was added to the "This is fun" paragraph.