- Fix: Calculator input box to start new line after calculation
The calculator input box was appending digits to the previous result after a calculation. This commit fixes this issue by adding a newLine variable that is set to true after a calculation. When a digit button is pressed, the input box is cleared and the new digit is displayed if newLine is true.